don't restart filebeat if it's not connected
If filebeat isn't "connected" (i.e. doesn't have an output configured),
trying to start it will fail, as it will error out with "beat.go:161: No
outputs are defined. Please define one under the output section." (at
least version 1.3, which the version in the default install_source)

Then, because its systemd unit file has "Restart=always", systemd will
try to restart it quickly a few times, and of course all attempts will
fail, so systemd will reach the restart threshold and stop trying the
restarts. It will also ignore any restart attempt for some time, and the
problem is that generally a legit restart request (i.e. with outputs
defined) is sent in that timeframe (because, for example, a relation to
logstash is finally established), and so it is ignored.

This results in filebeat not being running when a new unit has finished
running all its hooks, which is problematic. This patch should fix this
behavior.